feat: token refresh

+++ app/src/main/java/com/example/todomateclone/ui/diary/DiaryDetailFragment.kt
@@ -7,14 +7,13 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.inputmethod.InputMethodManager
-import androidx.core.view.GravityCompat
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import com.example.todomateclone.MainActivity
 import com.example.todomateclone.R
 import com.example.todomateclone.databinding.FragmentDiaryDetailBinding
-import com.example.todomateclone.viewmodel.CommentViewModel
 import com.example.todomateclone.viewmodel.DiaryViewModel
 import kotlinx.coroutines.launch
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -26,6 +25,14 @@ class DiaryDetailFragment : Fragment() {
     private val binding get() = _binding!!
     private val navigationArgs: DiaryDetailFragmentArgs by navArgs()
     private val diaryViewModel: DiaryViewModel by viewModel()
+
+    lateinit var mainActivity: MainActivity
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+
+        mainActivity = context as MainActivity
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -52,26 +59,40 @@ class DiaryDetailFragment : Fragment() {
             }
         }
 
-        toolbar.inflateMenu(R.menu.diary_menu)
-
-        // toolbar menu selected action
-        toolbar.setOnMenuItemClickListener {
-            when (it.itemId) {
-                R.id.edit_diary -> {
-                    val action = DiaryDetailFragmentDirections.actionDiaryDetailFragmentToDiaryEditFragment(navigationArgs.diaryId)
-                    this.findNavController().navigate(action)
-                    true
-                }
-                R.id.delete_diary -> {
-                    lifecycleScope.launch {
-                        diaryViewModel.deleteIdDiary(
-                            navigationArgs.diaryId
-                        )
-                        findNavController().navigateUp()
+        // 자신의 게시물일 때만 수정 및 삭제 가능
+        if (navigationArgs.searchDid == -1) {
+            // toolbar menu selected action
+            toolbar.inflateMenu(R.menu.diary_menu)
+            toolbar.setOnMenuItemClickListener {
+                when (it.itemId) {
+                    R.id.edit_diary -> {
+                        val action = DiaryDetailFragmentDirections.actionDiaryDetailFragmentToDiaryEditFragment(navigationArgs.diaryId)
+                        this.findNavController().navigate(action)
+                        true
                     }
-                    true
+                    R.id.delete_diary -> {
+                        AlertDialog.Builder(mainActivity)
+                            .setMessage("일기를 삭제하시겠습니까?")
+                            .setPositiveButton("예"
+                            ) { _, _ ->
+                                lifecycleScope.launch {
+                                    diaryViewModel.deleteIdDiary(
+                                        navigationArgs.diaryId
+                                    )
+                                    findNavController().navigateUp()
+                                }
+                                Log.d("MyTag", "positive")
+                            }
+                            .setNegativeButton("아니오"
+                            ) { _, _ ->
+                                Log.d("MyTag", "negative")
+                            }
+                            .create()
+                            .show()
+                        true
+                    }
+                    else -> true
                 }
-                else -> true
             }
         }
 
@@ -85,3 +106,4 @@ class DiaryDetailFragment : Fragment() {
         }
     }
 }
+
